added more functions
